+++ common/src/main/java/net/combatroll/mixin/EntityAttributesMixin.java
@@ -1,0 +1,19 @@
+package net.combatroll.mixin;
+
+import net.combatroll.CombatRoll;
+import net.combatroll.Platform;
+import net.minecraft.entity.attribute.EntityAttributes;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(EntityAttributes.class) // Low priority to be applied as last
+public class EntityAttributesMixin {
+    @Inject(method = "<clinit>", at = @At("TAIL"))
+    private static void static_tail_CombatRoll(CallbackInfo ci) {
+        if (Platform.Fabric) {
+            CombatRoll.registerAttributes();
+        }
+    }
+}